Revise tx completion stats.
This adds in an eligibility figure for all treatment-related tables.

It also revises eligibility to use the CDC's guidelines for time to
completion, which are longer than what I was using before.

Finally, it uses a more accurate calculation for the completion
denominator: people are counted if they have spent enough time on the
regimen OR if they've already completed. This is essential with the
longer CDC timelines, since many people will complete treatment right at
16 or 36 weeks, and should be counted right away - not 10-12 weeks later.
Put another way, we count completions as soon as they happen but allow
the full 16/36/etc. weeks before considering someone to be incomplete or
behind schedule.